Add 2-way communication between Javascript & bevy
We add a script to the webview that reads the webpage and sends a
request to bevy to spawn a sprite for each link in the page.

This is a direct copy `contributors.rs` with few changes.

An attempt was made to allow user to press space bar to navigate to the
currently highlighted web page, but wry seems to eat the input,
preventing the space bar from getting to bevy.

Limitations:

- The webview shows up, but is immediately overwritten by bevy. You can
  briefly see the webpage when moving around the window.
- The input is completely eaten by wry, preventing bevy interactions,
  and it also means you are clicking blindly on webpage links